Switched over to use the pappl-retrofit library
Most of the code of the PostScript Printer Application got valuable
for Printer Applications retro-fitting classic CUPS printer drivers in
general, as they also consist of PPD files and filters, which is what
the PostScript Printer Application supports.

After continuing the development of the code for a general
printer-driver-retro-fitting library and adding several new features
for the general use, it is now available as the libpappl-retrofit
library here:

https://github.com/OpenPrinting/pappl-retrofit

and future Printer Applications will use it.

The first in the series is the PostScript Printer Application which
with this commit is linking the library. So this repository only
contains a small code stub to configure the Printer Application for
PostScript printers, a test page, and snapcraft.yaml with some
additional files to build the Snap.

In the Snap the patch support-new-pappl-string-options.patch which was
formerly applied to the code of the PostScript Printer Application is
now moved over to the libpappl-retrofit library (updated to apply to
that code).

The Snap builds also libfontembed of cups-filters now, as the new
Printer Application did not start without the library.