Adapted to API change of pappl-retrofit
pappl-retrofit got new functionality for supporting multi-driver
Printer Applications, which needed an API change as there are more
configuration options for creating the Printer Application now.

See

https://github.com/OpenPrinting/pappl-retrofit/commit/a1b1aead73

The code of this Printer Application needed to get adapted to that.

Note that no extra functionality got added with this commit. The
behavior for the user is absolutely the same as before.