Makefile: Create alkl directories and links
When doing "make install" we now create all directories needed to
run the Printer Application. We also link our /usr/lib/ps-printer-app
directory to /usr/lib/cups if CUPS is installed. This way we can use filters
and backends installed for the system's CUPS.

Before we create the symlink to /usr/lib/cups, we remove the old
file/symlink/directory with a simple "rm" command and only if we
succeed we actually create the symlink. This way we do not remove a
directory nor do we create a symlink IN the destination directory
instead of TO the destination directory.