"*cupsFilter(2): ..." lines got removed from manufacturer PPDs in error
The global correction of common errors in the manufacturer-supplied,
ready-made PPD files also removed the "*cupsFilter(2): ..."
lines. This was wrong. All included PPD files are non PostScript and
need this line, even many PostScript PPDs need it.

This was probably a copy-and-paste error which slipped in when getting
this scriptlet line from some debian/rules file.

In addition, to improve convenience for developers (and everyone who
wants to build from source), we do a "make clean" before "make" in the
"ps-printer-app" part in snapcraft.yaml, because if we had done
"make" off-Snap, directly in the source tree. and afterwards build the
Snap with snapcraft, the build sucks in our local binary of
hplip-printer-app instead of compiling its own one in the Snap harness
with the appropriate libraries, ending up with the Snap containing an
executable which does not work inside the Snap. The "make clean"
removes any accidentally grabbed binary.

Also put some comments into snapcraft.yaml.