snapcraft.yaml: Remove libjpeg9 build from the Snap
Started off the project of retro-fitting Printer Applications from
hp-printer-app as example, the snapcraft.yaml does its own build of
libjpeg and here it uses libjpeg9 and not the libjpeg8 which comes
with Ubuntu by default. generally using newer libraries is desirable,
as they usually improve or fix something.

Here the problem is that the Snap Store's build servers have problems
to download the upstream tarball of this library and therefore onen
has to manually trigger each build several times until the build
servers succeeed to download the tarball and complete the build.

Therefore I tried to build the Snap with the libjpeg8 which comes with
Ubuntu Core 20, the underlying system for this Snap and it works
(including a photo print test where PAPPL opens a JPEG and converts it
to Raster). So I downgraded to Ubuntu's libjpeg8 which one simply
pulls in via "build-packages" and "stage-packages" in snapcraft.yaml.

This only made a (probably always having been there) bug in PAPPL
hitting, PAPPL's build system not including LDFLAGS when linking. So I
have patched this.

In addition, I also added some missing "cups" in the "after" entries
in snapcraft.yaml, to get a more reliable build of the Snap.