Auto-selection of printer model also if device ID has no "CMD:"
Auto-selection of the printer model when setting up a printer only
worked if there is a "CMD:" field in the device ID and PostScript was
listed as supported PDL.

Some backends do not obtain the device ID from the printer but obtain
make and model by another method and make an artificial device ID from
them, not having a "CMD:" field at all. If such a backend is selected via
the "Device" drop-down in the web interface, the auto-selection does not
work.

Now we accept both completely missing "CMD:" field or PostScript
listed, but not "CMD:" fields not listing PostScript.