Corrected passing of arguments in startup script for the Snap
The startup script scripts/run-*-printer-app has to pass on the
arguments it receives to the actual Printer Application executable,
exactly as it has received it, especially preserving spaces. This does
not work with $* but has to be "$@" instead. This way the command line
interface as designed for PAPPL also works with Printer Applications
in Snaps.

Now calls like

    /snap/bin/XXX-printer-app drivers -o 'device-id="MFG:HP;MDL:HP OfficeJet Pro 8730;CMD:PCL-XL;"'

actually work (this one lets the Printer Application tell whether and
how it supports the printer given by device ID).