In auto-add callback assign generic PPD only to PostScript printers
In the auto-add callback, ps_autoadd(), auto-assign the generic
PostScript printer PPD only to printers which actually support
PostScript and report this via CMD: entry in the device ID. This
especially prevents that this Printer Application reports any
arbitrary printer as supported.

If a device ID with an explicitly supported model name but without
CMD: is supplied, the printer is reported as supported as before.