Switched from HP's Upstream source to Debian's package source
As HPLIP has many bugs which got reported by distro maintainers but
not taken care of by HP's HPLIP developers, fixes got applied as
patches in the distro packages, ~80 (!) in the Debian package (which
is also used by Ubuntu).

Creating the Snap with the original source code from HP and not
applying all these patches would leave the HPLIP Printer Application
as buggy as the original HPLIP, but separately maintaining these
patches for the Snap, and especially if one tries to find out which
patches are the relevant ones for the Printer Application and which
not, would be a lot of unneeded work, also having to update the
patches when the Snap is updated to a new upstream source version of
HPLIP.

We use the Debian package's source code with all patches applied
already in the HPLIP and Ghostscript Printer Applications and now also
switch over here in the PostScript Printer Application, as some of the
patches also apply to the PPD files.

The code is downloaded from Debian's GitLab site:

https://salsa.debian.org/printing-team/hplip.v2/

Currently, to grab 3.21.12+dfsg0-1 the tag debian/3.21.12+dfsg0-1 is
downloaded.

Some changes in the build are done:

- In Debian's source the included PPD files are uncompressed for
  better GIT tracking, HP's original source has them gzipped. So we do
  not need to uncompress them.
- As we only need the hpps filter executable from the non-PPD files we
  do not run "./configure; make; make install" but compile this one
  executable by calling gcc directly (but we did this already before).